agregado referidor y monedas

+++ database/migrations/2017_04_18_214924_create_guias.php
@@ -16,7 +16,7 @@ class CreateGuias extends Migration
         Schema::create('guias', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('idiomas_id')->unsigned();
-            $table->float('costo');
+            $table->integer('costo');
             $table->timestamps();
 
             $table->foreign('idiomas_id')->references('id')->on('idiomas');
